chore: fix all clippy warnings (9 pre-existing)
- serialization.rs: redundant  cast
- alsa.rs: clamp pattern, channels.get(0) -> first()
- pipewire.rs: channels.get(0) -> first(), redundant u32 cast,
  loop variable indexing -> enumerate
- jack.rs: redundant closure -> fn ptr
- adrift/runtime: redundant closures -> fn ptr

+++ rill-graph/src/serialization.rs
@@ -244,7 +244,7 @@ impl GraphDocument {
         Self {
             format_version: "rill/1".to_string(),
             sample_rate,
-            block_size: B as usize,
+            block_size: B,
             resources,
             nodes,
             connections,
